fix: change prodict Price and SalePrice data types from decimal to double

+++ ECommerce.Core/Domain/Entities/Product.cs
@@ -18,9 +18,9 @@ namespace ECommerce.Core.Domain.Entities
         public string? Description { get; set; }
         [Required]
         [Range(0, double.MaxValue)]
-        public decimal Price { get; set; }
+        public double Price { get; set; }
         [Range(0, double.MaxValue)]
-        public decimal? SalePrice { get; set; }
+        public double? SalePrice { get; set; }
         public string? ImageUrl { get; set; }
         [Required]
         public long Stock { get; set; }
